requirements.txt
gif==23.3.0
joblib==1.4.2
keras==3.5.0
matplotlib==3.8.4
meegkit==0.1.9
mne==1.9.0
mne_bids==0.16.0
mne_icalabel==0.6.0
numpy==2.2.3
pandas==2.2.3
Pillow==11.1.0
plotly==5.22.0
pre-commit
pyctf==1.0
pyprep==0.4.3
scikit_learn==1.4.2
scipy==1.15.2
seaborn==0.13.2
tensorflow==2.17.0
tqdm==4.66.4
tqdm_joblib==0.0.4
wasabi==1.1.3
